Creates two popup html pages
