+++ chrome_extension/haskeys_popup.html
@@ -1,11 +1,12 @@
 <html>
   <head>
-    <script src="./extension_popup.js" type="module"></script>
     <link rel="stylesheet" type="text/css" href="./styles/popup_stylesheet.css">
   </head>
   <body>
     <div id="popup-container">
       <h1>Social Sign Extension</h1>
+      <h2>Getting started!</h2>
+      <p>Documentation here ...</p>
     </div>
   </body>
 </html>
