+++ chrome_extension/manifest.json
@@ -10,7 +10,7 @@
         "128": "images/icon-128.png"
     },
     "action": {
-      "default_popup": "extension_popup.html"
+      "default_popup": "nokeys_popup.html"
     },
     "content_scripts": [
         {
