+++ chrome_extension/nokeys_popup.html
@@ -1,0 +1,19 @@
+<html>
+  <head>
+    <script src="./nokeys_popup.js" type="module"></script>
+    <link rel="stylesheet" type="text/css" href="./styles/popup_stylesheet.css">
+  </head>
+  <body>
+    <div id="popup-container">
+      <h1>Social Sign Extension</h1>
+      <button id="generation-btn">Generate Key-pair</button>
+      <form>
+        <label for="update-code">
+          Enter Update Code:
+          <input type="text" id="update-code" name="update-code">
+        </label>
+        <input type="submit" value="Update Key-pair">
+      </form>
+    </div>
+  </body>
+</html>
