+++ chrome_extension/styles/popup_stylesheet.css
@@ -1,7 +1,3 @@
 body{
   width: 250px;
 }
-
-div {
-  text-align: center;
-}
